scripts/embeddings/pyproject.toml
[project]
name = "embedding-umap"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "umap-learn>=0.5.5",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]
